--- Depricated/OldHome2.html
@@ -1,79 +1,0 @@
-<div class="BGWrapper">
-
-    <div class="MountWrapper">
-        <img class="Mount" src="Images/snowMountFullRes.jpg">
-        <div class="MountTitle">
-            <h1 class="JDName">Jeremy DeFossett</h1>
-            <h1 class="JDComp">Developer</h1>
-        </div>
-    </div>
-
-    <div class="IntroWrapper">
-        <h1 class="IntroHead">Introduction</h1>
-        <div class="SuitPic">
-            <img style="width: 100%;" src="Images/SuitPort.jpg">
-        </div>
-        <div class="IntroTxt">
-            <p>	Why-How-What.  Simon Sinek stated that great leaders start with why, rather than what.  Why am I here?  I want to empower people with great technologies
-                in order for them to achieve their goals and desires.  How do I accomplish this?  By working to the best of my ability, constantly learn new technologies, and
-                bettering myself on a daily basis.  So what do I do?  I'm a developer.  I attend Indiana University pursuing a bachelors degree in Computer Science, a certificate
-                in Informatics, and a minor in business.  I don't plan on stopping there, however.  After graduating, I will be working to earn my Masters in Information Systems,
-                and later down the road my Masters in Business Administration.  I want to continue to work my way up, as I do not believe in complacency, while giving back to
-                those around me.</p>
-
-                <h2>"When you want to succeed, as bad as you want to breathe, then you'll be successful"</h2>
-
-                <a href="../Downloadables/Resume-DeFossett.pdf" download="Resume-DeFossett.pdf"><button type="button" class="btn btn-primary btn-lg JDResumeButton">Résumé</button></a>
-
-        </div>
-    </div>
-
-    <div class="PortfolioWrapper">
-        <h1 class="PortHead">Portfolio</h1>
-
-        <a href="https://github.com/jdeffo/JDeffo-Web"><div class="PortSamp1">
-            <img style="width: 100%;" src="Images/stockCode1.jpeg">
-            <h3>deffoWeb</h3>
-        </div></a>
-
-        <a href="https://github.com/jdeffo/deffoStocks"><div class="PortSamp2">
-            <img style="width: 100%;" src="Images/stockCode2.jpeg">
-            <h3>deffoStocks</h3>
-        </div></a>
-
-        <a href="https://github.com/jdeffo/To-Do-List"><div class="PortSamp3">
-            <img style="width: 100%;" src="Images/stockCode3.jpeg">
-            <h3>To-Do</h3>
-        </div></a>
-
-        <div style="display: table; width: 0; height: 0; clear: both; visibility: none;"></div>
-        <div class="PortGit">
-            <h2>The rest of my projects can be viewed on Github.com/JDeffo</h2>
-        </div>
-    </div>
-
-    <div class="ContactWrapper">
-        <h1 style="position: relative; left: 55px; top: 20px;">Contact</h1>
-
-        <div class="ContLinkedin">
-            <a href="https://linkedin.com/in/jdefossett">
-              <img style="width: 100%;" src="Images/JDLinked.png"></a>
-        </div>
-
-        <div class="ContGit">
-            <a href="https://github.com/jdeffo">
-              <img style="width: 100%;" src="Images/JDGithub.png"></a>
-        </div>
-
-        <div class="ContEmail" style="text-align: center;">
-            <a href ng-click="hctrl.controller.showContactModal()">
-              <img style="width: 100%;" src="Images/JDEmail.png"></a>
-              <h5>Click to Send Me a Message</h5>
-        </div>
-        <div style="display: table; width: 0; height: 0; clear: both;"></div>
-    </div>
-
-    <div class="GraphicsWrapper">
-        <h3>Graphics design by <a href="https://www.linkedin.com/in/heathbauman/" style="color: #000000;">Heath Bauman</a></h3>
-    </div>
-</div>
