--- Depricated/OldStyles2.css
@@ -1,141 +1,0 @@
-
-/*~~~~~NAV~~~~~*/
-.JDLogo {
-  width: 50%;
-  margin-top: -7%;
-}
-
-/*~~~~~HOME PAGE~~~~~*/
-
-/*WRAPPERS*/
-.BGWrapper {
-    background-color: #cecece;
-    padding-bottom: 1%;
-    font-family: Bebas Neue;
-    overflow: hidden;
-}
-.MountWrapper {
-    margin-top: 2.6%;
-    height: 480px;
-    width: 100%;
-}
-.IntroWrapper {
-    height: 375px;
-    margin-top: -1.25%;
-    width: 100%;
-}
-.PortfolioWrapper {
-    height: 550px;
-    background: #b5b5ce;
-    width: 100%;
-}
-.ContactWrapper {
-    height: 300px;
-    width: 100%;
-    margin-top: -1.2%;
-}
-.GraphicsWrapper {
-    width: 100%;
-    margin-top: -2%;
-    text-align: center;
-}
-/*CONTENT*/
-
-/*MountWrapper*/
-.Mount {
-    width: 100%;
-}
-.MountTitle {
-    text-align: center;
-    margin-top: -7%;
-}
-/*IntroWrapper*/
-.SuitPic {
-    width: 267px;
-    height: 272px;
-    margin-left: 3%;
-    margin-top: 1%;
-    float: left;
-}
-.IntroTxt {
-  font-size: 170%;
-  overflow: hidden;
-  text-align: center;
-  padding: 20px;
-  /*padding-right: 322px;*/
-}
-.IntroHead {
-    position: relative;
-    left: 55px;
-    top: 20px;
-}
-/*PortfolioWrapper*/
-.PortHead {
-    position: relative;
-    left: 55px;
-    top: 20px;
-}
-.PortSamp1 {
-    /*width: 535px;*/
-    width: 28%;
-    /*height: 410px;*/
-    float: left;
-    margin-left: 3%;
-    margin-top: 1%;
-    text-align: center;
-    color: black;
-}
-.PortSamp2 {
-    width: 28%;
-    /*width: 535px;
-    height: 410px;*/
-    float: left;
-    margin-left: 4.5%;
-    margin-top: 1%;
-    text-align: center;
-    color: black;
-}
-.PortSamp3 {
-    width: 28%;
-    /*width: 535px;
-    height: 410px;*/
-    float: left;
-    margin-left: 4.5%;
-    margin-top: 1%;
-    text-align: center;
-    color: black;
-}
-.PortGit {
-    text-align: center;
-}
-/*ContactWrapper*/
-.ContLinkedin {
-  position: relative;
-  width: 15%;
-  margin-left: 13.75%;
-  margin-top: 2.5%;
-  float: left;
-}
-
-.ContGit {
-  position: relative;
-  width: 15%;
-  margin-left: 13.75%;
-  margin-top: 1.5%;
-  float: left;
-}
-
-.ContEmail {
-  position: relative;
-  width: 12%;
-  margin-left: 16.75%;
-  margin-top: 1.5%;
-  float: left;
-}
-
-/*~~~~~~Modal~~~~~~~*/
-
-/*Invalid*/
-input.ng-invalid.ng-touched, textarea.ng-invalid.ng-touched{
-    border: 2px solid red;
-}
