--- Depricated/Refactor1/Old.html
@@ -1,83 +1,0 @@
-<div class="JDBgWrapper">
-  <img class="JDMount" src="Images/snowMountFullRes.jpg">
-
-  <div class="JDTitleWrapper">
-      <h1 class="JDName">Jeremy DeFossett</h1>
-      <h1 class="JDComp">Developer</h1>
-  </div>
-
-    <div class="JDIntroWrapper">
-      <h1 style="font-family: Bebas Neue; margin-left: 3%; margin-top: 1.5%;">Introduction</h1>
-
-      <img class="JDSelfPort" src="Images/SuitPort.jpg">
-
-      <div class="JDIntroTxt">
-        <p>	Why-How-What.  Simon Sinek stated that great leaders start with why, rather than what.  Why am I here?  I want to empower people with great technologies
-        in order for them to achieve their goals and desires.  How do I accomplish this?  By working to the best of my ability, constantly learn new technologies, and
-        bettering myself on a daily basis.  So what do I do?  I'm a developer.  I attend Indiana University pursuing a bachelors degree in Computer Science, a certificate
-        in Informatics, and a minor in business.  I don't plan on stopping there, however.  After graduating, I will be working to earn my Masters in Information Systems,
-        and later down the road my Masters in Business Administration.  I want to continue to work my way up, as I do not believe in complacency, while giving back to
-        those around me.</p>
-
-        <h2 style="text-align: center; margin-left: -15%;">"When you want to succeed, as bad as you want to breathe, then you'll be successful"</h2>
-
-        <a href="../Downloadables/Resume-DeFossett.pdf" download="Resume-DeFossett.pdf"><button type="button" class="btn btn-primary btn-lg JDResumeButton">Résumé</button></a>
-      </div>
-    </div>
-
-      <div class="JDPortfolioWrapper">
-
-        <h1 style="font-family: Bebas Neue; margin-left: 3%; margin-top: 1.5%;">Portfolio</h1>
-
-        <div class="JDCodeSampleContainer">
-
-          <a href="https://github.com/jdeffo/deffoStocks"><div class="JDSamp1">
-            <img style="width: 100%;" src="Images/stockCode1.jpeg">
-              <h3>deffoStocks</h3>
-          </div></a>
-
-          <a href="https://github.com/jdeffo/GTD"><div class="JDSamp2">
-              <img style="width: 100%;" src="Images/stockCode2.jpeg">
-              <h3>GTD</h3>
-          </div><a>
-
-          <a href="https://github.com/jdeffo/To-Do-List"><div class="JDSamp3">
-              <img style="width: 100%;" src="Images/stockCode3.jpeg">
-              <h3>To-Do</h3>
-          </div></a>
-
-          <div style="display: table; width: 0; height: 0; clear: both;">
-
-          </div>
-        </div>
-
-        <h2 style="text-align:center;">The rest of my projects can be viewed on Github.com/JDeffo</h2>
-
-      </div>
-
-      <div class="JDContactWrapper">
-        <h1 style="font-family: Bebas Neue; margin-left: 3%; magin-top: 1.5%;">Contact</h1>
-
-      <div class="JDLinkedinContact">
-        <a href="https://linkedin.com/in/jdefossett">
-          <img style="width: 100%;" src="Images/JDLinked.png"></a>
-      </div>
-
-      <div class="JDGithubContact">
-        <a href="https://github.com/jdeffo">
-          <img style="width: 100%;" src="Images/JDGithub.png"></a>
-      </div>
-
-      <div class="JDEmailContact">
-        <a href="https://github.com/jdeffo">
-          <img style="width: 100%;" src="Images/JDEmail.png"></a>
-      </div>
-
-      <div style="display: table; width: 0; height: 0; clear: both;"></div>
-     </div>
-
-    <div class="creds" style="text-align:center;">
-      <br><br><br>
-      <h3 style="font-family: Bebas Neue;">Graphics design by <a href="https://www.linkedin.com/in/heathbauman/" style="color: #000000;">Heath Bauman</a></h3>
-    </div>
-</div>
