--- Depricated/Refactor1/OldStyles.css
@@ -1,171 +1,0 @@
-/*Style Sheet for Website
-Organized by page
-Alphabetical
-*/
-
-/*--INDEX--*/
-.JDBackground {
-  background-color: #cecece;
-}
-
-/*--NAVBAR-*/
-/*.JDNavWrapper {
-    padding-bottom: 0;
-    display: block;
-}*/
-
-.JDNavText {
-  font-family: Bebas Neue;
-}
-
-.JDLogo {
-  width: 50%;
-  margin-top: -7%;
-}
-
-/*--HOME--*/
-.JDBgWrapper {
-  width: 100%;
-  z-index: 0;
-  vertical-align: top;
-  padding-top: 0;
-  margin-top: -1%;
-  position: relative;
-}
-
-.JDHomeAbout {
-  border-bottom: 1px solid black;
-  background-color: #636f77;
-  margin-top: -2%;
-}
-
-.JDMount {
-  margin-top: 0;
-  width: 100%;
-  /*opacity: 50%;*/
-}
-
-.JDTitleWrapper {
-  background-color: e9e9f4;
-  text-align: center;
-  margin-top: -7%;
-}
-
-.JDName {
-  font-family: Bebas Neue;
-  font-size: 400%;
-}
-
-.JDComp {
-  font-family: Bebas Neue;
-  margin-top: 0;
-}
-
-.JDIntroWrapper {
-  height: 350px;
-}
-
-.JDSelfPort {
-    width: 14%;
-    margin-left: 3%;
-    float: left;
-}
-
-.JDIntroTxt {
-  font-family: Bebas Neue;
-  font-size: 150%;
-  float: left;
-  margin-left: 19%;
-  margin-top: -14.5%;
-  padding-right: 2%;
-}
-
-.JDButtonHold {
-  position: relative;
-  width: 86.5%;
-  padding: 0;
-  text-align: center;
-  margin-left: 50%;
-  margin-right: 50%;
-}
-
-.JDResumeButton {
-  position: relative;
-  margin-left: -125%;
-  background-color: #c2b0d7;
-  color: black;
-  border-color: #c2b0d7;
-}
-
-.JDPortfolioWrapper {
-  background-color: #b5b5ce;
-  height: 540px;
-  font-family: Bebas Neue;
-}
-
-.JDCodeSampleContainer {
-  width: 100%;
-}
-
-.JDProjLine {
-  width: 100%;
-
-}
-
-.blank {
-  float: left;
-  visibility: hidden;
-}
-
-.JDSamp1 {
-  position: relative;
-  width: 28%;
-  margin-left: 4%;
-  float: left;
-  text-align: center;
-  color: black;
-}
-
-.JDSamp2 {
-  position: relative;
-  width: 28%;
-  margin-left: 4%;
-  float: left;
-  text-align: center;
-  color: black;
-}
-
-.JDSamp3 {
-  position: relative;
-  width: 28%;
-  margin-left: 4%;
-  text-align: center;
-  color: black;
-  float: left;
-}
-
-.JDContactWrapper {
-    height: 200px;
-}
-
-.JDLinkedinContact {
-  position: relative;
-  width: 15%;
-  margin-left: 13.75%;
-  margin-top: 1.5%;
-  float: left;
-}
-
-.JDGithubContact {
-  position: relative;
-  width: 15%;
-  margin-left: 13.75%;
-  float: left;
-}
-
-.JDEmailContact {
-  position: relative;
-  width: 12%;
-  margin-left: 16.75%;
-  float: left;
-}
