--- Depricated/Refactor2/About.html
@@ -1,24 +1,0 @@
-<div class="SuitPic" style="margin-top: 2.5%;">
-    <img class="img-circle" src="Images/SuitPort.jpg">
-</div>
-<div class="IntroPageText">
-  <p>As you can tell, I'm a developer.  I have mostly back-end and web development experience, but also have
-    strong interests in machine learning and mobile development.  I am currently learning technologies
-    involved with machine learning, such as tensorFlow, as well as iOS and Android development.
-  </p>
-  <p>I have more interests than just development though.  I have a strong interest in business, and how
-    business relates to technology.  I think the connection between business and technology is one of
-    the most important aspects of a successful company.  I also follow the stock market daily; I love
-    investing and the importance of setting a stable future for my family.
-  </p>
-  <p>The most important aspect of my life is my family, and especially my parents.  Everything I have
-  and will accomplish is for them.  I also enjoy spending time with my friends, exercising, playing
-  sports, snowboarding out west, and traveling.
-  </p>
-</div>
-
-<hr />
-
-<div class="IntroPageResume">
-  <h2>My Resume</h3>
-</div>
