--- Depricated/Refactor2/ContactModal.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" ng-app="JDeffoApp">
-  <head>
-    <title>Jeremy DeFossett</title>
-    <link href="Content/Bootstrap/css/bootstrap.min.css" rel="stylesheet" type="text/css" />
-    <link href="Content/DeffoStyles.css" rel="stylesheet" type="text/css" />
-    <link rel="shortcut icon" href="Images/JDIcon.ico" />
-    <script src="app/lib/angular.min.js"></script>
-    <script src="app/lib/angular-route.min.js"></script>
-    <script src="app/moduleConfig.js"></script>
-  </head>
-  <body>
-
-      <div class="modal-header">
-          <h3 class="modal-title">Contact Me</h3>
-      </div>
-      <div class="modal-body contactModal" style="overflow: auto;">
-
-          <form name="contactForm" ng-submit="hmctrl.sendMessage()" novalidate class="form-group">
-              <input type="text" placeholder="Name" name="name" class="form-control" ng-model="hmctrl.model.contact.name" ng-required="true"/><br>
-              <input type ="email" placeholder="Email" name="email" class="form-control" ng-model="hmctrl.model.contact.email" ng-required="true"/><br>
-              <textarea placeholder="Your Message" name="message" class="form-control" ng-model="hmctrl.model.contact.message" ng-required="true"></textarea><br>
-              <input type="submit" value="Send" class="btn btn-primary pull-right" ng-disabled="contactForm.$invalid"/>
-          </form>
-        {{hmctrl.model.contact.name}}
-      </div>
-      <div class="modal-footer">
-          <button class="btn btn-primary" type="button" ng-click="hmctrl.close()">Close</button>
-      </div>
-  </body>
-  </html>
