--- Depricated/Refactor2/DeffoStyles.css
@@ -1,166 +1,0 @@
-/*~~~~~NAVBAR~~~~~*/
-.JDLogo {
-  width: 50%;
-  margin-top: -7%;
-}
-/*~~~HOME~~~*/
-
-/*--Wrappers--*/
-.MountWrapper{
-    width:100%;
-    height:45%;
-    position: relative;
-}
-.IntroWrapper {
-    position: relative;
-    padding-top: .5%;
-}
-.PortfolioWrapper {
-    background: #b5b5ce;
-    position: relative;
-    padding-top: .5%;
-    padding-bottom: .5%;
-}
-.ContactWrapper {
-    position: relative;
-    padding-top: .5%;
-    padding-bottom: .5%;
-}
-.indent {
-    margin-left: 2%;
-}
-/*--Content--*/
-
-/*Mountain Stock Photo*/
-.MountTitle{
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    text-align: center;
-}
-.MountPic {
-    width: 100%;
-    height: 100%;
-}
-/*Intro Secction*/
-.SuitPic{
-    width: 14%;
-    margin-left: auto;
-    margin-right: auto;
-}
-.img-circle {
-    border-radius: 50%;
-    width: 100%;
-}
-.IntroTxt {
-    width: 98%;
-    margin-left: auto;
-    margin-right: auto;
-    position: relative;
-    text-align: center;
-}
-.IntroTxt p {
-    font-size: 150%;
-}
-.QuoteTxt {
-    text-align: center;
-    width: 100%;
-}
-.QuoteTxt p {
-    font-size: 140%;
-}
-.resumeButton {
-    width: 100%;
-    height: 10%;
-    padding-top: 0;
-    text-align: center;
-}
-/*Portfolio Section*/
-.PortSamp {
-    width:30%;
-    margin-left: 2.5%;
-    float: left;
-    text-align: center;
-}
-.PortGitTxt {
-    width: 100%;
-    white-space: nowrap;
-    text-align: center;
-}
-.PortGitTxt h2 {
-    bottom: 0;
-}
-/*Contact Section*/
-.ContactLinkedin{
-    width:15%;
-    margin-left: 13.75%;
-    float: left;
-}
-.ContactGit{
-    width:15%;
-    margin-left: 13.75%;
-    float: left;
-    text-align: center;
-}
-.ContactEmail{
-    width:15%;
-    margin-left: 13.75%;
-    float: left;
-    text-align: center;
-}
-/*~~~About Page~~~*/
-.IntroPageText {
-  width: 75%;
-  text-align: center;
-  font-size: 150%;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 2.5%;
-}
-.IntroPageResume {
-  text-align: center;
-}
-/*~~~Portfolio Page~~~*/
-.ProjectsContainer {
-  text-align: center;
-  position: relative;
-}
-.ProjWrap {
-  width:30%;
-  margin-left: 2.5%;
-  margin-top: 2.5%;
-  float: left;
-  text-align: center;
-}
-
-/*~~~Contact Page~~~*/
-.ContactInfo {
-  text-align: center;
-}
-.ContactInfo p {
-  margin-left: auto;
-  margin-right: auto;
-  text-align: center;
-}
-.ContactForm {
-  width: 60%;
-  text-align: center;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-/*~~~HTML Elements~~~*/
-
-footer {
-  text-align: center;
-}
-html,body{
-    height:100%;
-    padding-top: 25px;
-    /*font-family: Bebas Neue;*/
-    background-color: #cecece;
-    margin: 0;
-}
-/*::-webkit-scrollbar {
-    display: none;
-}*/
