--- Depricated/Refactor2/Home.html
@@ -1,87 +1,0 @@
-<!--Mount Wrapepr-->
-<div class="MountWrapper">
-    <img class="MountPic" src="Images/snowMountFullRes.jpg">
-    <div class="MountTitle">
-        <h1 class="JDName">Jeremy DeFossett</h1>
-        <h2 class="JDComp">Developer</h2>
-    </div>
-</div>
-<!--Intro Wrapper-->
-<div class="IntroWrapper">
-    <div class="SuitPic">
-        <img class="img-circle" src="Images/SuitPort.jpg">
-    </div>
-
-    <div class="IntroTxt">
-        <p>Why-How-What.  Simon Sinek stated that great leaders start with why, rather than what.  Why am I here?  I want to empower people with great technologies
-        in order for them to achieve their goals and desires.  How do I accomplish this?  By working to the best of my ability, constantly learn new technologies, and
-        bettering myself on a daily basis.  So what do I do?  I'm a developer.  I attend Indiana University pursuing a bachelors degree in Computer Science, a certificate
-        in Informatics, and a minor in business.  I don't plan on stopping there, however.  After graduating, I will be working to earn my Masters in Information Systems,
-        and later down the road my Masters in Business Administration.  I want to continue to work my way up, as I do not believe in complacency, while giving back to
-        those around me.</p>
-    </div>
-
-    <div class="QuoteTxt">
-        <p>"When you want to succeed, as bad as you want to breathe, then you'll be successful"</p>
-    </div>
-
-    <div class="resumeButton">
-        <a href="../Downloadables/Resume-DeFossett.pdf" download="Resume-DeFossett.pdf"><button type="button" class="btn btn-primary btn-lg JDResumeButton">Résumé</button></a>
-    </div>
-
-</div>
-<!--Portfolio Wrapper-->
-<div class="PortfolioWrapper">
-    <h1 class="indent">Portfolio</h1>
-
-    <div class="PortSamp">
-        <a href="https://github.com/jdeffo/JDeffo-Web">
-            <img style="width: 100%;" src="Images/stockCode1.jpeg">
-        </a>
-        <h3>deffoWeb</h3>
-    </div>
-
-    <div class="PortSamp">
-        <a href="https://github.com/jdeffo/deffoStocks">
-            <img style="width: 100%;" src="Images/stockCode2.jpeg">
-        </a>
-        <h3>deffoStocks</h3>
-    </div>
-
-    <div class="PortSamp">
-        <a href="https://github.com/jdeffo/To-Do-List">
-            <img style="width: 100%;" src="Images/stockCode3.jpeg">
-        </a>
-        <h3>To-Do</h3>
-    </div>
-    <div style="display: table; width: 0; height: 0; clear: both; visibility: none;"></div>
-
-    <div class="PortGitTxt">
-        <h2>The rest of my projects can be viewed on <a href="https://github.com/jdeffo">github.com/jdeffo</a></h2>
-    </div>
-
-</div>
-<!--Contact Wrapper-->
-<div class="ContactWrapper">
-    <h1 class="indent">Contact</h1>
-
-    <div class="ContactLinkedin">
-        <a href="https://linkedin.com/in/jdefossett">
-          <img style="width: 100%; height: 65%;" src="Images/JDLinked.png"></a>
-    </div>
-
-    <div class="ContactGit">
-        <a href="https://github.com/jdeffo">
-          <img style="width: 75%; height: 60%;" src="Images/JDGithub.png"></a>
-    </div>
-
-    <div class="ContactEmail">
-        <a href ng-click="hctrl.controller.showContactModal()">
-          <img style="width: 75%; height: 60%;" src="Images/JDEmail.png"></a>
-    </div>
-<div style="display: table; width: 0; height: 0; clear: both; visibility: none;"></div>
-</div>
-
-<footer>
-  <h2>Graphic Design by <a href="https://www.linkedin.com/in/heathbauman/">Heath Bauman</a></h2>
-</footer>
