--- Depricated/Refactor2/NavBar.html
@@ -1,17 +1,0 @@
-<div class="JDNavWrapper">
-  <nav class="navbar navbar-inverse navbar-fixed-top" style="border-style: none;">
-    <div class="container-fluid" style="background-color: #888789;">
-      <div class="navbar-header">
-        <a class="navbar-brand" href="#" style="color: black;"><img class="JDLogo" src="Images/JDLogo.png"></a>
-      </div>
-
-      <div class="collapse navbar-collapse">
-        <ul class="nav navbar-nav pull-right JDNavText">
-          <li><a href="#!about" style="color: black; font-size: 140%;">About</a></li>
-          <li><a href="#!portfolio" style="color: black; font-size: 140%;">Portfolio</a></li>
-          <li><a href="#!contact" style="color: black; font-size: 140%;">Contact</a></li>
-        </ul>
-      </div>
-    </div>
-  </nav>
-</div>
