--- Depricated/Refactor2/Portfolio.html
@@ -1,18 +1,0 @@
-<div class="ProjectsContainer">
-
-  <h1>My personal projects</h1>
-
-  <hr />
-
-  <div class="ProjWrap" ng-repeat="card in pctrl.controller.cardList">
-      <a href="{{card.link}}">
-          <img style="width: 100%;" src="{{card.image}}">
-      </a>
-      <h3>{{card.title}}</h3>
-  </div>
-
-
-  <div style="display: table; width: 0; height: 0; clear: both; visibility: none;"></div>
-
-
-</div>
