+++ index.html
@@ -3,22 +3,24 @@
   <head>
     <title>Jeremy DeFossett</title>
     <link href="Content/Bootstrap/css/bootstrap.min.css" rel="stylesheet" type="text/css" />
-    <link href="Content/DeffoStyles.css" rel="stylesheet" type="text/css" />
+    <link href="Content/Portfolio-Styles.css" rel="stylesheet" type="text/css" />
     <link href="Content/Responsive.css" rel="stylesheet" type="text/css" />
-    <link rel="shortcut icon" href="Images/deffoLogo.ico" />
+    <link href="Content/css/font-awesome.min.css" rel="stylesheet" type="text/css" />
+    <link rel="shortcut icon" href="Images/Portfolio-Deffo-Logo.ico" />
     <script src="app\lib\angular.min.js"></script>
     <script src="app\lib\angular-route.min.js"></script>
     <script src="app\lib\ui-bootstrap-tpls-2.5.0.min.js"></script>
     <script src="app/lib/jquery-3.2.1.min.js"></script>
     <script src="app\app.js"></script>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+    <!--Keyword Tags-->
+    <meta name="description" content="Portfolio website for Jeremy DeFossett; Consultant, Investor, Developer.">
+    <meta name="keywords" content="Consultant,Investor,Developer,Software,Engineering,Web,
+    Development,Portfolio,HTML,CSS,JavaScript,Consulting,Investing,Computer,Science">
+    <meta name="author" content="Jeremy DeFossett">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-
   </head>
 
   <body ng-app="JDeffoApp">
-
-  <!--<header ng-include="'Views/NavBar.html'" ng-controller="HomeController"></header>-->
 
     <main ng-view style="height: 100%; "></main>
